pyproject.toml
[project]
name = "xgov-flask-admin"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "email-validator>=2.0.0",
    "flask-admin>=2.0.0a0",
    "flask>=3",
    "govuk-frontend-jinja==3.8.0",
    "govuk-frontend-wtf==3.2.0",
    "sqlalchemy>2",
    "tox-uv>=1.29.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.0",
    "faker>=37.4.2",
    "flask-sqlalchemy-lite>=0.1.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-flask>=1.3.0",
    "pytest-playwright>=0.7.1",
    "requests>=2.32.5",
    "ruff>=0.14.0",
    "tox>=4.23.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
ignore-vcs = false

[tool.hatch.build.targets.wheel]
packages = ["src/xgov_flask_admin"]
force-include = { "src/xgov_flask_admin/static" = "xgov_flask_admin/static" }

[tool.hatch.build.targets.sdist]
include = [
    "src/",
]
force-include = { "src/xgov_flask_admin/static" = "src/xgov_flask_admin/static" }

[tool.uv]

[tool.uv.sources]
flask-admin = { git = "https://github.com/flask-admin/flask-admin", branch = "master" }

[tool.coverage.run]
source = ["src/xgov_flask_admin"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.pytest.ini_options]
filterwarnings = [
    "error",
]
